Add files via upload
With this analysis we seek to determine who was the best driver of the year 2016 in order to reward him with a bonus for the work done. To do this, we decided to study different KPIs: Rating, amount of income, number of miles traveled, number of trips made and average time of each trip per driver.